refactor: move each shell layer into distinct directories
