+++ src/shell/lexer/handler/quote_handler.ts
