+++ src/shell/lexer/handler/word_handler.ts
