+++ src/shell/lexer/lexer.ts
